Adding auto-fix buildifier and auto-fix rubocop
  * to bin/pre-commit file which is symlinked by bin/setup
    to .git/hooks/pre-commit.

If, after running the formatters it's determined that some files have
changed (using git status -s), it shows the diff between the two states
and aborts the commit so that you have a chance to add fixed files.

It does not automatically add anything to git because it would be
a very large and often invalid assumption.